+++ data-raw/0 - create all.R
@@ -7,18 +7,16 @@ files <- c(
   "Comorbidity_diagn_codes.R",
   "Comorbidity_weights.R",
   "I_APME_DEM_AUTOR_CRITR_ETEN_CM.R",
-  # "internal_datas.R",
+  "internal_datas.R",
   "Obstetrics_diagn_codes.R",
-  # "Pop_QC.R",
-  # "RLS_convert.R",
-  # "RLS_list.R",
-  "V_DEM_PAIMT_MED_CM.SMED_COD_DIN.R",
-  "V_DEM_PAIMT_MED_CM.SMED_COD_SERV.R",
+  "Pop_QC.R",
+  "RLS_convert.R",
+  "RLS_list.R",
+  "V_DEM_PAIMT_MED_CM.R",
   "V_DENOM_COMNE_MED.R",
   "V_DES_COD.R",
-  "V_PRODU_MED.NOM_MARQ_COMRC.R"
+  "V_PRODU_MED.R"
 )
-# 22 mins
 files <- paste0("data-raw/", files)
 
 t1 <- Sys.time()